BUILD: Defer loading of Markdown-rendering pipeline until needed.
