+++ packages/frontend/src/llm_conversation/conversation_editor.tsx
@@ -1,10 +1,6 @@
 import { createForm, getValue, reset, type SubmitHandler } from "@modular-forms/solid";
 import Send from "lucide-solid/icons/send";
-import rehypeKatex from "rehype-katex";
-import remarkGfm from "remark-gfm";
-import remarkMath from "remark-math";
-import { For, Match, Switch, Show } from "solid-js";
-import { SolidMarkdown } from "solid-markdown";
+import { For, lazy, Match, Suspense, Switch, Show } from "solid-js";
 
 import { LLMInteraction } from "catcolab-document-types";
 import { Button } from "catcolab-ui-components";
@@ -22,6 +18,8 @@ export function LLMConversationEditor(props: {
     conversation: ApiLLMConversation;
     documentStore: ApiDocumentStore;
 }) {
+    void MarkdownMessage.preload();
+
     const inferenceKey = useInferenceKey();
     const controller = createLLMConversationController(
         () => props.conversation,
@@ -44,15 +42,17 @@ export function LLMConversationEditor(props: {
     return (
         <div class="llm-conversation">
             <div class="llm-transcript">
-                <For each={props.conversation.interactions()}>
-                    {(interaction) => <LLMInteractionView interaction={interaction} />}
-                </For>
-                <For each={controller.state.liveInteractions}>
-                    {(interaction) => <LLMInteractionView interaction={interaction} />}
-                </For>
-                <Show when={controller.state.streamingContent}>
-                    <MarkdownMessage content={controller.state.streamingContent} />
-                </Show>
+                <Suspense>
+                    <For each={props.conversation.interactions()}>
+                        {(interaction) => <LLMInteractionView interaction={interaction} />}
+                    </For>
+                    <For each={controller.state.liveInteractions}>
+                        {(interaction) => <LLMInteractionView interaction={interaction} />}
+                    </For>
+                    <Show when={controller.state.streamingContent}>
+                        <MarkdownMessage content={controller.state.streamingContent} />
+                    </Show>
+                </Suspense>
             </div>
             <Form onSubmit={onSubmit}>
                 <Field name="message">
@@ -89,18 +89,4 @@ export const LLMInteractionView = (props: { interaction: LLMInteraction }) => (
     </Switch>
 );
 
-/** Render Markdown content, such as a message from the LLM. */
-export const MarkdownMessage = (props: { content: string }) => (
-    <SolidMarkdown
-        class="markdown-message"
-        renderingStrategy="reconcile"
-        remarkPlugins={remarkPlugins}
-        rehypePlugins={rehypePlugins}
-    >
-        {props.content}
-    </SolidMarkdown>
-);
-
-/** Support GitHub-flavored Markdown, plus math as rendered in MDX help pages. */
-const remarkPlugins = [remarkGfm, remarkMath];
-const rehypePlugins = [rehypeKatex];
+const MarkdownMessage = lazy(() => import("./markdown_message"));
